Remove nesting from grammar productions which end with a unicode prime
The process of transforming left-recursive rules into non-left-recursive
forms also changes the shape of the resulting parse-tree. This change
allows you to  undo some of those changes. If the names of the
"synthetic" rules added during  the transformation end with a unicode
prime character then they will not be nested.